rename cu_dynthreshold_impl
