+++ components/tasks/cu_dynthreshold/src/lib.rs
@@ -2,4 +2,4 @@
 mod cu_dynthreadold_impl;
 
 #[cfg(feature = "gst")]
-pub use cu_dynthreadold_impl::*;
+pub use cu_dynthreshold_impl::*;
